api/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Naked==0.1.31
pycryptodome==3.10.1
PyYAML==5.4.1
requests==2.26.0
shellescape==3.8.1
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
Flask-Admin==1.5.8
Flask-BasicAuth==0.2.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1

